backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
itsdangerous==2.2.0
groq==0.13.0
PyPDF2==3.0.1
pdfplumber==0.11.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.4.0
psycopg2-binary==2.9.9
cloudinary==1.41.0
gunicorn==22.0.0
requests==2.32.3
httpx==0.27.2
